Fixed tooltip offset bug
The tooltip was being offset by the top value of #Map, which was pushing
it off the screen. It turns out that #Map's top value was at the bottom
of the piki image, thereby making the offset value very large. The top
value of the .piki table was used instead with success.