Added function to tabs
Tabs are now functional. When a tab is clicked, it will change the piki
shapes that are present on the parent piki. The view page renders these
new piki shapes from a "dense" pikiChildren array (called
pikiChildren_dense). This is done because the pikiChildren normally is
quite sparse, and the ng-repeat will create a div for all empty items.

Minor:
+Loads all piki children at once
+Added alpha = 0.6 to the piki shapes
+Changed ng-repeat by adding "track by $index". Otherwise it would throw
an error because it would see the multiple "Piki" objects as duplicates.