Made drawing tool factories and added circle tool
Create factory objects for drawing on the link canvas. Each factory is
for a specific tool. E.g., freehand, circle, etc. The factories share a
set of drawing functions for their specific shape. The LinkCtrl stores
which tool is active, and calls the common drawing functions when the
drawing directive triggers a certain action.

The circle tool was added as well.