Fixed bug that occurred after drawing one shape
There was a bug that occurred only after the user drew one shape. This
was occurring because the xyArray would have a null object that the draw
methods were attempting to draw.

Additionally, a "resetTool" function was added to each drawing tool that
is activated after the drawingTool factory detects that drawing is
completed.