Fixed bug: Rectangle shapes were not drawing correctly
After drawing a rectangle shape, it would disappear. Furthermore,
drawing multiple shapes wouldn't lead to multiple shapes being
displayed. This had to do with the draw method of the
drawingRectangleTool clearing the canvas each time a new rectangle was
drawn.

Furthermore, a bug was fixed where the xyArray would have a "undefined"
object within it. This was fixed by checking if the xyArray's current
object was undefined when the mouse moved and making it an empty array
if it was.